+++ Src/Core/VpnHood.Core.Client.Device.Android/AndroidVpnService.cs
@@ -17,9 +17,9 @@ namespace VpnHood.Core.Client.Device.Droid;
 [Service(
     Permission = Manifest.Permission.BindVpnService,
     Exported = false,
-// #if !DEBUG 
+#if !DEBUG  //todo
     Process = ProcessName,
-//#endif
+#endif
     ForegroundServiceType = ForegroundService.TypeSystemExempted)]
 [IntentFilter(["android.net.VpnService"])]
 public class AndroidVpnService : VpnService, IVpnServiceHandler