+++ Tests/VpnHood.AppLib.Test/Tests/AccessCodeTest.cs
@@ -1,5 +1,4 @@
-﻿using System.Net.Sockets;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Text;
 using VpnHood.AppLib.ClientProfiles;
 using VpnHood.Core.Common.Exceptions;